Temporary Wall/Ceiling/Floor Modular Set
Made a temporary modular set for the walls, ceilings and floors. Hopefully I can figure out the texture issues later because that is such a big issue. If that was fixed, the level can be made VERY quickly. I made this set in about 5-10 minutes (after trying to fix the texture issues for a few hours, but this temporary solution will have to work for now).